==> notebook/06_visualisations.ipynb <==
Import Statements

%load_ext autoreload
%autoreload 2

import sys
import os
import pandas as pd
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

# Add parent directory to sys.path
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), "..")))

from src.config import TRANSFORMED_DATA_DIR
from src.plot_utils import plot_aggregated_time_series



Loading merged parquet file and seperating the attributes and target

# Load the tabular data
features_and_targets = pd.read_parquet(TRANSFORMED_DATA_DIR / "tabular_data.parquet")
features = features_and_targets.iloc[:, :-1]
targets = features_and_targets["target"]
features.head(5)
features.shape
targets.head(5)
targets.shape

# Select a specific row to visualize (e.g., row 3772)
# You can change the row_id to visualize a different station or time
row_id = 3772
features[features['start_station_name'] == features.iloc[row_id]['start_station_name']]



# Plot the time series for the selected row
plot_aggregated_time_series(features, targets, row_id=row_id)
